% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predhy.GUI.R
\name{predhy.GUI}
\alias{predhy.GUI}
\title{Graphical User Interface for R package predhy}
\usage{
predhy.GUI()
}
\value{
No return value, called for Graphical User Interface
}
\description{
Graphical User Interface for cross validation, genotype conversion and hybrid performance prediction.
}
\examples{
{
predhy.GUI()}
}
